requirements.txt
# SkillBridge Bot — Python Dependencies
# Install with:  pip install -r requirements.txt

# Telegram Bot framework (async, v3.x)
aiogram==3.7.0

# Modern async HTTP client used internally by aiogram
aiohttp==3.9.5

# Web framework and server
fastapi==0.111.0
uvicorn==0.30.1

# Database ORM (SQLAlchemy wrapper)
sqlmodel==0.0.19
psycopg2-binary==2.9.9

# Environment variable management
python-dotenv==1.0.1
pydantic-settings==2.3.1

# Rate limiting and API protection
slowapi==0.1.9

# Performance and caching
redis==5.0.1

# Authentication & JWT
PyJWT==2.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart>=0.0.7

# Database migrations
alembic==1.13.1

# Jinja2 (required by FastAPI for templates/static)
jinja2>=3.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
httpx==0.25.2
